#272 Move Log to its own component called LogViewerList
