+++ src/js/view/components/LogViewer.js
@@ -1,85 +1,20 @@
 import React from 'react';
 import {
-  LogViewContainer,
-  CloseFileButton,
-  Log,
-  LogLine
+  LogViewerContainer,
+  CloseFileButton
 } from '../styledComponents/LogViewerStyledComponents';
+import LogViewerList from './LogViewerList';
 import { connect } from 'react-redux';
 import { closeFile } from './helpers/handleFileHelper';
-import {
-  calculateSize,
-  calculateWrap,
-  maxLength
-} from 'js/view/components/helpers/measureHelper';
-import TextHighlightRegex from './TextHighlightRegex';
-import WindowedList from 'react-list';
-import {
-  filterByRegExp,
-  parseRegExp
-} from 'js/view/components/helpers/regexHelper.js';
+import { parseRegExp } from 'js/view/components/helpers/regexHelper.js';
 
 class LogViewer extends React.Component {
   constructor(props) {
     super(props);
-    this.logRef = React.createRef();
-    this.windowedListRef = React.createRef();
-    this.rulerRef = React.createRef();
     this.state = {
       escapeRegexSequence: '@'
     };
   }
-
-  scrollToBottom = el => {
-    el.scrollAround(this.lastIndex);
-  };
-
-  componentDidUpdate() {
-    if (this.props.tailSwitch) {
-      // this.scrollToBottom(this.logRef.current);
-      this.scrollToBottom(this.windowedListRef.current);
-    }
-  }
-
-  updateSizes() {}
-
-  itemRenderer = (lineWidth, lines, regex) => {
-    return (i, key) => {
-      return (
-        <LogLine
-          key={key}
-          index={i}
-          fixedWidth={lineWidth}
-          wrap={this.props.wrapLineOn ? 'true' : undefined}
-        >
-          {regex && regex.test(lines[i], regex) ? (
-            <TextHighlightRegex
-              text={lines[i]}
-              color={this.props.highlightColor}
-              regex={regex}
-            />
-          ) : (
-            lines[i]
-          )}
-        </LogLine>
-      );
-    };
-  };
-
-  itemSizeGetter = (lines, charSize, clientWidth) => {
-    const sizes = lines.map(line => {
-      return calculateWrap(charSize, line, clientWidth);
-    });
-    return index => {
-      return sizes[index];
-    };
-  };
-
-  constItemSizeGetter = size => {
-    return index => {
-      return size;
-    };
-  };
 
   render() {
     const highlightRegex = parseRegExp(
@@ -91,27 +26,8 @@ class LogViewer extends React.Component {
       this.state.escapeRegexSequence
     );
 
-    const lines = filterByRegExp(this.props.liveLines, filterRegex);
-
-    this.lastIndex = lines.length - 1;
-
-    const charSize = this.rulerRef.current
-      ? calculateSize('a', this.rulerRef.current)
-      : [0, 0];
-
-    const lineWidth =
-      this.logRef.current && this.props.wrapLineOn
-        ? this.logRef.current.clientWidth
-        : maxLength(lines) * charSize[1];
-
-    const itemSizeGetter = this.props.wrapLineOn
-      ? this.itemSizeGetter(lines, charSize, this.logRef.current.clientWidth)
-      : this.constItemSizeGetter(charSize[0]);
-
-    const itemRenderer = this.itemRenderer(lineWidth, lines, highlightRegex);
-
     return (
-      <LogViewContainer>
+      <LogViewerContainer>
         <CloseFileButton
           openFiles={this.props.openFiles}
           onClick={() => {
@@ -121,24 +37,15 @@ class LogViewer extends React.Component {
             );
           }}
         />
-        <Log ref={this.logRef}>
-          <LogLine
-            style={{
-              visibility: 'hidden',
-              minWidth: 0,
-              display: 'inline-block'
-            }}
-            ref={this.rulerRef}
-          />
-          <WindowedList
-            ref={this.windowedListRef}
-            itemRenderer={itemRenderer}
-            itemSizeGetter={itemSizeGetter}
-            length={lines.length}
-            type="variable"
-          />
-        </Log>
-      </LogViewContainer>
+        <LogViewerList
+          highlightColor={this.props.highlightColor}
+          wrapLines={this.props.wrapLineOn}
+          scrollToBottom={this.props.tailSwitch}
+          lines={this.props.liveLines}
+          highlightRegex={highlightRegex}
+          filterRegex={filterRegex}
+        />
+      </LogViewerContainer>
     );
   }
 }
